requirements.txt
# DeepAgent Financial Systems Requirements

# Core LangChain and LangGraph
langchain>=0.3.20
langchain-openai==0.2.8
langgraph>=0.6.7
langchain-core>=0.3.17
langchain-community==0.3.7
langgraph-cli[inmem]>=0.4.0
openai>=1.0.0

# Web Search and Research
tavily-python==0.5.0

# Financial Data
yfinance==0.2.40
pandas==2.2.3
numpy==1.26.4

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
typing-extensions==4.12.2

# Development and Testing
jupyter==1.1.1
notebook==7.2.2
ipykernel==6.29.5

# Optional: For enhanced financial analysis
matplotlib==3.9.2
plotly==5.24.1
scipy==1.14.1
scikit-learn==1.5.2

# Async support
aiohttp>=3.11.14
asyncio

# Data processing
requests==2.32.3
beautifulsoup4==4.12.3